requirements.txt
tensorflow
streamlit
Pillow
numpy
openai